Move registry to its own file
